Check the variable balance between the probability and non-probability samples

Check the variable balance between the probability and non-probability samples

Usage

check_balance(x, object, dig)

Arguments

x

Formula specifying variables to check

object

Object of nonprobsvy class

dig

Number of digits for rounding (default = 2)

Value

A list containing nonprobability totals, probability totals, and their differences

Examples

data(admin)
data(jvs)

jvs_svy <- svydesign(ids = ~ 1,  weights = ~ weight,
strata = ~ size + nace + region, data = jvs)

ipw_est1 <- nonprob(selection = ~ region + private + nace + size,
target = ~ single_shift,
svydesign = jvs_svy,
data = admin, method_selection = "logit"
)

ipw_est2 <- nonprob(
selection = ~ region + private + nace + size,
target = ~ single_shift,
svydesign = jvs_svy,
data = admin, method_selection = "logit",
control_selection = control_sel(h = 1, est_method_sel = "gee"))

## check the balance for the standard IPW
check_balance(~size, ipw_est1)
#> $nonprob_totals
#>     sizeL     sizeM     sizeS 
#>  8193.385 13529.601 31175.142 
#> 
#> $prob_totals
#> sizeL sizeM sizeS 
#>  8561 13758 29551 
#> 
#> $balance
#>   sizeL   sizeM   sizeS 
#> -367.62 -228.40 1624.14 
#> 

## check the balance for the calibrated IPW
check_balance(~size, ipw_est2)
#> $nonprob_totals
#> sizeL sizeM sizeS 
#>  8561 13758 29551 
#> 
#> $prob_totals
#> sizeL sizeM sizeS 
#>  8561 13758 29551 
#> 
#> $balance
#> sizeL sizeM sizeS 
#>     0     0     0 
#>