Control parameters for inference

controlInf constructs a list with all necessary control parameters for statistical inference.

Usage

controlInf(
  vars_selection = FALSE,
  var_method = c("analytic", "bootstrap"),
  rep_type = c("auto", "JK1", "JKn", "BRR", "bootstrap", "subbootstrap", "mrbbootstrap",
    "Fay"),
  bias_inf = c("union", "div"),
  num_boot = 500,
  bias_correction = FALSE,
  alpha = 0.05,
  cores = 1,
  keep_boot,
  pmm_exact_se = FALSE,
  pi_ij
)

Arguments

vars_selection

If TRUE, then variables selection model is used.

var_method

variance method.

rep_type

replication type for weights in the bootstrap method for variance estimation passed to survey::as.svrepdesign(). Default is subbootstrap.

bias_inf

inference method in the bias minimization.

if union then final model is fitting on union of selected variables for selection and outcome models

if div then final model is fitting separately on division of selected variables into relevant ones for selection and outcome model.

num_boot

number of iteration for bootstrap algorithms.

bias_correction

if TRUE, then bias minimization estimation used during fitting the model.

alpha

Significance level, Default is 0.05.

cores

Number of cores in parallel computing.

keep_boot

Logical indicating whether statistics from bootstrap should be kept. By default set to TRUE

pmm_exact_se

Logical value indicating whether to compute the exact standard error estimate for pmm estimator. The variance estimator for estimation based on pmm can be decomposed into three parts, with the third being computed using covariance between imputed values for units in probability sample using predictive matches from non-probability sample. In most situations this term is negligible and is very computationally expensive so by default this is set to FALSE, but it is recommended to set this value to TRUE before submitting final results.

pi_ij

TODO, either matrix or ppsmat class object.

Value

List with selected parameters.

See also