Control parameters for outcome model

controlOut constructs a list with all necessary control parameters for outcome model.

Usage

controlOut(
  epsilon = 1e-04,
  maxit = 100,
  trace = FALSE,
  k = 1,
  penalty = c("SCAD", "lasso", "MCP"),
  a_SCAD = 3.7,
  a_MCP = 3,
  lambda_min = 0.001,
  nlambda = 100,
  nfolds = 10,
  treetype = "kd",
  searchtype = "standard",
  predictive_match = 1:2,
  pmm_weights = c("none", "prop_dist"),
  pmm_k_choice = c("none", "min_var"),
  pmm_reg_engine = c("glm", "loess")
)

Arguments

epsilon

Tolerance for fitting algorithms. Default is 1e-6.

maxit

Maximum number of iterations.

trace

logical value. If TRUE trace steps of the fitting algorithms. Default is FALSE.

k

The k parameter in the RANN::nn2() function. Default is 5.

penalty

penalty algorithm for variable selection. Default is SCAD

a_SCAD

The tuning parameter of the SCAD penalty for outcome model. Default is 3.7.

a_MCP

The tuning parameter of the MCP penalty for outcome model. Default is 3.

lambda_min

The smallest value for lambda, as a fraction of lambda.max. Default is .001.

nlambda

The number of lambda values. Default is 100.

nfolds

The number of folds during cross-validation for variables selection model.

treetype

Type of tree for nearest neighbour imputation passed to RANN::nn2() function.

searchtype

Type of search for nearest neighbour imputation passed to RANN::nn2() function.

predictive_match

(Only for predictive mean matching) Indicates how to select 'closest' unit from nonprobability sample for each unit in probability sample. Either 1 (default) or 2 where 1 is matching by minimizing distance between \(\hat{y}_{i}\) for \(i \in S_{A}\) and \(y_{j}\) for \(j \in S_{B}\) and 2 is matching by minimizing distance between \(\hat{y}_{i}\) for \(i \in S_{A}\) and \(\hat{y}_{i}\) for \(i \in S_{A}\).

pmm_weights

(Only for predictive mean matching) Indicate how to weight k nearest neighbours in \(S_{B}\) to create imputed value for units in \(S_{A}\). The default value "none" indicates that mean of k nearest \(y\)'s from \(S_{B}\) should be used whereas "prop_dist" results in weighted mean of these k values where weights are inversely proportional to distance between matched values.

pmm_k_choice

Character value indicating how k hyper-parameter should be chosen, by default "none" meaning k provided in control_outcome argument will be used. For now the only other option "min_var" means that k will be chosen by minimizing estimated variance of estimator for mean. Parameter k provided in this control list will be chosen as starting point.

pmm_reg_engine

TODO

Value

List with selected parameters.

See also