Control parameters for selection model

controlSel constructs a list with all necessary control parameters for selection model.

Usage

controlSel(
  method = "glm.fit",
  epsilon = 1e-04,
  maxit = 500,
  trace = FALSE,
  optimizer = c("maxLik", "optim"),
  maxLik_method = "NR",
  optim_method = "BFGS",
  dependence = FALSE,
  key = NULL,
  est_method_sel = c("mle", "gee"),
  h = c(1, 2),
  penalty = c("SCAD", "lasso", "MCP"),
  a_SCAD = 3.7,
  a_MCP = 3,
  lambda = -1,
  lambda_min = 0.001,
  nlambda = 50,
  nfolds = 10,
  print_level = 0,
  start_type = c("glm", "naive")
)

Arguments

method

estimation method.

epsilon

Tolerance for fitting algorithms by default 1e-6.

maxit

Maximum number of iterations.

trace

logical value. If TRUE trace steps of the fitting algorithms. Default is FALSE

optimizer

optimization function for maximum likelihood estimation.

maxLik_method

maximisation method that will be passed to maxLik::maxLik() function. Default is NR.

optim_method

maximisation method that will be passed to stats::optim() function. Default is BFGS.

dependence

logical value - TRUE if samples are dependent.

key

binary key variable

est_method_sel

Method of estimation for propensity score model.

h

Smooth function for the generalized estimating equations methods taking the following values

if 1 then \(\mathbf{h}\left(\mathbf{x}, \boldsymbol{\theta}\right) = \frac{\pi(\mathbf{x}, \boldsymbol{\theta})}{\mathbf{x}}\)

if 2 then \( \mathbf{h}\left(\mathbf{x}, \boldsymbol{\theta}\right) = \mathbf{x}\)

penalty

The penanlization function used during variables selection.

a_SCAD

The tuning parameter of the SCAD penalty for selection model. Default is 3.7.

a_MCP

The tuning parameter of the MCP penalty for selection model. Default is 3.

lambda

A user-specified \(\lambda\) value during variable selection model fitting.

lambda_min

The smallest value for lambda, as a fraction of lambda.max. Default is .001.

nlambda

The number of lambda values. Default is 50.

nfolds

The number of folds for cross validation. Default is 10.

print_level

this argument determines the level of printing which is done during the optimization (for propensity score model) process.

start_type

Type of method for start points for model fitting taking the following values

if glm then start taken from the glm function called on samples.

if naive then start consists of a vector which has the value of an estimated parameter for one-dimensional data (on intercept) and 0 for the rest.

Value

List with selected parameters.

See also

Author

Łukasz Chrostowski, Maciej Beręsewicz