Control parameters for the selection model

control_sel constructs a list with all necessary control parameters for selection model.

Usage

control_sel(
  est_method = c("mle", "gee"),
  gee_h_fun = 1,
  optimizer = c("maxLik", "optim"),
  maxlik_method = "NR",
  optim_method = "BFGS",
  epsilon = 1e-04,
  maxit = 500,
  trace = FALSE,
  penalty = c("SCAD", "lasso", "MCP"),
  a_SCAD = 3.7,
  a_MCP = 3,
  lambda = -1,
  lambda_min = 0.001,
  nlambda = 50,
  nfolds = 10,
  print_level = 0,
  start_type = c("glm", "naive", "zero"),
  nleqslv_method = c("Broyden", "Newton"),
  nleqslv_global = c("dbldog", "pwldog", "cline", "qline", "gline", "hook", "none"),
  nleqslv_xscalm = c("fixed", "auto"),
  dependence = FALSE,
  key = NULL
)

Arguments

est_method

Method of estimation for propensity score model.

gee_h_fun

Smooth function for the generalized estimating equations (GEE) method taking the following values

if 1 then h(x, ) = (x, )x

if 2 then h(x, ) = x

optimizer

optimization function for maximum likelihood estimation.

maxlik_method

maximisation method that will be passed to maxLik::maxLik() function. Default is NR.

optim_method

maximisation method that will be passed to stats::optim() function. Default is BFGS.

epsilon

Tolerance for fitting algorithms by default 1e-6.

maxit

Maximum number of iterations.

trace

logical value. If TRUE trace steps of the fitting algorithms. Default is FALSE

penalty

The penalization function used during variables selection.

a_SCAD

The tuning parameter of the SCAD penalty for selection model. Default is 3.7.

a_MCP

The tuning parameter of the MCP penalty for selection model. Default is 3.

lambda

A user-specified value during variable selection model fitting.

lambda_min

The smallest value for lambda, as a fraction of lambda.max. Default is .001.

nlambda

The number of lambda values. Default is 50.

nfolds

The number of folds for cross validation. Default is 10.

print_level

this argument determines the level of printing which is done during the optimization (for propensity score model) process.

start_type

Type of method for start points for model fitting taking the following values

if glm then start taken from the glm function called on samples.

if naive then start consists of a vector which has the value of an estimated parameter for one-dimensional data (on intercept) and 0 for the rest.

if zero then start is a vector of zeros.

nleqslv_method

The method that will be passed to nleqslv::nleqslv() function.

nleqslv_global

The global strategy that will be passed to nleqslv::nleqslv() function.

nleqslv_xscalm

The type of x scaling that will be passed to nleqslv::nleqslv() function.

dependence

logical value (default TRUE) informing whether samples overlap (NOT YET IMPLEMENTED, FOR FUTURE DEVELOPMENT).

key

binary key variable allowing to identify the overlap (NOT YET IMPLEMENTED, FOR FUTURE DEVELOPMENT).

Value

List with selected parameters.

See also