Summary statistics for model of nonprobsvy class.

Summary statistics for model of nonprobsvy class.

Usage

# S3 method for class 'nonprobsvy'
summary(object, test = c("t", "z"), correlation = FALSE, cov = NULL, ...)

Arguments

object

object of nonprobsvy class

test

Type of test for significance of parameters "t" for t-test and "z" for normal approximation of students t distribution, by default "z" is used if there are more than 30 degrees of freedom and "t" is used in other cases.

correlation

correlation Logical value indicating whether correlation matrix should be computed from covariance matrix by default FALSE.

cov

Covariance matrix corresponding to regression parameters

...

Additional optional arguments

Value

An object of summary_nonprobsvy class containing:

call – A call which created object.

pop_total – A list containing information about the estimated population mean, its standard error and confidence interval.

sample_size – The size of the samples used in the model.

population_size – The estimated size of the population from which the non–probability sample was drawn.

test – Type of statistical test performed.

control – A List of control parameters used in fitting the model.

model – A descriptive name of the model used, e.g., "Doubly-Robust", "Inverse probability weighted", or "Mass Imputation".

aic – Akaike's information criterion.

bic – Bayesian (Schwarz's) information criterion.

residuals – Residuals from the model, providing information on the difference between observed and predicted values.

likelihood – Logarithm of likelihood function evaluated at coefficients.

df_residual – Residual degrees of freedom.

weights – Distribution of estimated weights obtained from the model.

coef – Regression coefficients estimated by the model.

std_err – Standard errors of the regression coefficients.

w_val – Wald statistic values for the significance testing of coefficients.

p_values – P-values corresponding to the Wald statistic values, assessing the significance of coefficients.

crr – The correlation matrix of the model coefficients, if requested.

confidence_interval_coef – Confidence intervals for the model coefficients.

names – Names of the fitted models.